pyproject.toml
[project]
name = "create-images-dataset"
version = "0.1.0"
description = "Audio to spatial features image converter"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.16.1",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "librosa>=0.10.0",
    "spaudiopy>=0.2.0",
    "nnAudio>=0.3.1",
    "soundfile>=0.12.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "tifffile>=2023.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "scikit-image>=0.21.0",
    "tqdm>=4.65.0",
    "joblib>=1.3.0",
    "h5py>=3.9.0"
]
